Moving IndexedDigitalSurface to topology directory/package
